27th commit: 정근, 연장, 야간, 휴일 급여계산식 로직 구현

+++ WebApp0840/src/main/java/cz/yeo/wage/WebApp0840/app/user/dto/UserContext.java
@@ -15,7 +15,7 @@ public class UserContext extends User {
     private final Integer id;
     private final String nickname;
     private final LocalDateTime createDate;
-    private final Double BaseWage;
+    private final Integer BaseWage;
     private final Integer annual;
     private final String userImgUrl;
